Add built branch that only contains dist files.
We need to patch the plugin for our use case quite urgently. We have a
way of patching plugin files from zip files, so easist way to grab a
custom zip file for this plugin is to create a built branch that
contains the files only needed for the plugin. We can then download a
zip of this from github and download into our WP project.